msm8916-common: Enable low RAM flag
With Android 10, RAM usage can be close to 1gb even without installing GApps and other applications so for Osprey the low ram flag should now be enabled even on 2gb models.

The low ram flag also disables some graphical features which improves system responsiveness.
With this flag and others already added, we are now close to the 'Android Go' configuration.